Update quantity to long

+++ app/src/main/java/edu/karolinawidz/homestocktracker/data/local/Item.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity
 data class Item(
     @PrimaryKey val name: String,
-    val quantity: Int,
+    val quantity: Long,
     val category: String
 )
